Добавил эндпоинты, подправил структуру пакетов в контроллерах и сервисах, добавил @Transactional к метода SubscriptionService
