+++ src/main/java/school/faang/user_service/controller/event/EventControllerValidation.java
@@ -7,10 +7,10 @@ import school.faang.user_service.exception.DataValidationException;
 
 import java.time.LocalDateTime;
 
-import static school.faang.user_service.exception.ExceptionMessage.INVALID_EVENT_END_DATE_EXCEPTION;
-import static school.faang.user_service.exception.ExceptionMessage.INVALID_EVENT_START_DATE_EXCEPTION;
-import static school.faang.user_service.exception.ExceptionMessage.NULL_EVENT_OWNER_ID_EXCEPTION;
-import static school.faang.user_service.exception.ExceptionMessage.NULL_OR_BLANK_EVENT_TITLE_EXCEPTION;
+import static school.faang.user_service.exception.message.ExceptionMessage.INVALID_EVENT_END_DATE_EXCEPTION;
+import static school.faang.user_service.exception.message.ExceptionMessage.INVALID_EVENT_START_DATE_EXCEPTION;
+import static school.faang.user_service.exception.message.ExceptionMessage.NULL_EVENT_OWNER_ID_EXCEPTION;
+import static school.faang.user_service.exception.message.ExceptionMessage.NULL_OR_BLANK_EVENT_TITLE_EXCEPTION;
 
 @Component
 class EventControllerValidation {
