+++ src/main/java/school/faang/user_service/exception/message/ExceptionMessage.java
@@ -1,4 +1,4 @@
-package school.faang.user_service.exception;
+package school.faang.user_service.exception.message;
 
 public enum ExceptionMessage {
     NULL_OR_BLANK_EVENT_TITLE_EXCEPTION("The event cannot have null-valued or blank title."),
