+++ src/main/java/school/faang/user_service/exception/message/MessageForGoalInvitationService.java
@@ -1,4 +1,4 @@
-package school.faang.user_service.exception;
+package school.faang.user_service.exception.message;
 
 import lombok.Getter;
