+++ src/main/java/school/faang/user_service/service/goal/GoalInvitationServiceValidator.java
@@ -1,4 +1,4 @@
-package school.faang.user_service.service;
+package school.faang.user_service.service.goal;
 
 import org.springframework.stereotype.Component;
 import school.faang.user_service.dto.goal.GoalInvitationDto;
@@ -8,7 +8,11 @@ import school.faang.user_service.exception.DataValidationException;
 
 import java.util.List;
 
-import static school.faang.user_service.exception.MessageForGoalInvitationService.*;
+import static school.faang.user_service.exception.message.MessageForGoalInvitationService.INVITED_HAS_GOAL;
+import static school.faang.user_service.exception.message.MessageForGoalInvitationService.INVITER_ID_EQUALS_INVITED_USER_ID;
+import static school.faang.user_service.exception.message.MessageForGoalInvitationService.MORE_THEN_THREE_GOALS;
+import static school.faang.user_service.exception.message.MessageForGoalInvitationService.NO_INVITED_IN_GOAL_INVITATION;
+import static school.faang.user_service.exception.message.MessageForGoalInvitationService.SET_GOALS_IS_NULL;
 
 @Component
 public class GoalInvitationServiceValidator {
