+++ src/main/java/school/faang/user_service/service/goal/filter/InvitedIdFilter.java
@@ -1,4 +1,4 @@
-package school.faang.user_service.service.filter;
+package school.faang.user_service.service.goal.filter;
 
 import org.springframework.stereotype.Component;
 import school.faang.user_service.dto.goal.InvitationFilterDto;
@@ -7,14 +7,14 @@ import school.faang.user_service.entity.goal.GoalInvitation;
 import java.util.stream.Stream;
 
 @Component
-public class InvitedNamePatternFilter implements InvitationFilter {
+public class InvitedIdFilter implements InvitationFilter {
     @Override
     public boolean isApplicable(InvitationFilterDto filters) {
-        return filters.getInvitedNamePattern() != null;
+        return filters.getInvitedId() != null;
     }
 
     @Override
     public Stream<GoalInvitation> apply(GoalInvitation invitation, InvitationFilterDto filters) {
-        return Stream.of(invitation).filter(goalInvitation -> goalInvitation.getInvited().getUsername().equals(filters.getInvitedNamePattern()));
+        return Stream.of(invitation).filter(goalInvitation -> goalInvitation.getInvited().getId() == filters.getInvitedId());
     }
 }
