+++ src/test/java/school/faang/user_service/controller/GoalInvitationControllerTest.java
@@ -11,7 +11,7 @@ import school.faang.user_service.controller.goal.GoalInvitationController;
 import school.faang.user_service.dto.goal.GoalInvitationDto;
 import school.faang.user_service.dto.goal.InvitationFilterDto;
 import school.faang.user_service.entity.RequestStatus;
-import school.faang.user_service.service.GoalInvitationService;
+import school.faang.user_service.service.goal.GoalInvitationService;
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.mockito.Mockito.verify;
